fix: clippy — byte_table above the test module, drop a redundant cast
Both slipped into the previous commit because the lint output was being
swallowed by a shell fallback. Checked properly now: clippy -D warnings
is clean across core, infer and the CLI.

+++ crates/needle-cli/src/main.rs
@@ -182,7 +182,7 @@ fn run_v3(o: &Opts, model: &str) {
             .max_tokens
             .unwrap_or(needle_infer::v3_engine::DEFAULT_MAX_NEW_TOKENS),
         temperature: o.temperature.unwrap_or(0.0),
-        seed: o.seed.unwrap_or(0) as u64,
+        seed: o.seed.unwrap_or(0),
         system: o.system.clone(),
         constrain: o.constrain,
     };
